pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssa-urban-deprivation-benchmark"
version = "0.1.0"
description = "Open multimodal pipeline for urban deprivation measurement in sub-Saharan Africa."
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.10",
  "numpy>=2.0",
  "pandas>=3.0",
  "PyYAML>=6.0.3",
  "requests>=2.33",
  "scikit-learn>=1.8",
]

[project.optional-dependencies]
geo = [
  "geopandas>=1.1",
  "osmnx>=2.1",
  "pyogrio>=0.11",
  "pyproj>=3.7",
  "rasterio>=1.5",
  "rasterstats>=0.20",
  "shapely>=2.1",
]
spatial = [
  "esda>=2.9",
  "libpysal>=4.14",
]
ml = [
  "torch>=2.11",
  "torchvision>=0.26",
  "xgboost>=3.2",
]
dev = [
  "pytest>=9.0",
]

[project.scripts]
udb = "ssa_urban_deprivation_benchmark.cli:main"
uda = "ssa_urban_deprivation_benchmark.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
